+++ CHANGELOG.md
@@ -8,15 +8,16 @@
   - This should fix the `java.lang.AssertionError: Class with incorrect id found` exception when using `kotlinx.serialization`
 - Fall back to `findTask` if `assembleProvider` of AndroidVariant is null when hooking source bundle and native symbols upload tasks ([#639](https://github.com/getsentry/sentry-android-gradle-plugin/pull/639))
 - Hook source context tasks to also run after `install{Variant}` tasks ([#643](https://github.com/getsentry/sentry-android-gradle-plugin/pull/643))
+- Do not run sentry-cli commands if telemetry is disabled ([#648](https://github.com/getsentry/sentry-android-gradle-plugin/pull/648))
 - Proguard and source context tasks don't run on every build ([#634](https://github.com/getsentry/sentry-android-gradle-plugin/pull/634))
   - Proguard UUID task now depends on the proguard mapping file. I.e. it will only run if the mapping file has changed
   - Source context tasks now depend on source file changes, if there are no source changes, the tasks won't run
 
 ### Dependencies
 
-- Bump CLI from v2.25.0 to v2.26.0 ([#638](https://github.com/getsentry/sentry-android-gradle-plugin/pull/638), [#640](https://github.com/getsentry/sentry-android-gradle-plugin/pull/640), [#642](https://github.com/getsentry/sentry-android-gradle-plugin/pull/642))
-  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2260)
-  - [diff](https://github.com/getsentry/sentry-cli/compare/2.25.0...2.26.0)
+- Bump CLI from v2.25.0 to v2.27.0 ([#638](https://github.com/getsentry/sentry-android-gradle-plugin/pull/638), [#640](https://github.com/getsentry/sentry-android-gradle-plugin/pull/640), [#642](https://github.com/getsentry/sentry-android-gradle-plugin/pull/642), [#647](https://github.com/getsentry/sentry-android-gradle-plugin/pull/647))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2270)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.25.0...2.27.0)
   Bump Android SDK from v7.2.0 to v7.3.0 ([#646](https://github.com/getsentry/sentry-android-gradle-plugin/pull/646))
     [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#730)
     [diff](https://github.com/getsentry/sentry-java/compare/7.2.0...7.3.0)